Make remote file syncing apply only to datasets
